requirements.txt
Django>=4.2
scikit-learn>=1.2
pandas>=2.0
numpy>=1.25
joblib>=1.3
djangorestframework>=3.15
python-dotenv>=1.0
gunicorn